requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pandas
Pillow
plotly
kaleido==0.2.1
Jinja2
WeasyPrint
reportlab
